pyproject.toml
[tool.poetry]
name = "booking-app"
version = "0.1.0"
description = ""
authors = ["Xzenoff <xzenonoff@gmail.com>"]
readme = "README.md"
packages = [{include = "booking_app"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0"
djangorestframework = "^3.14.0"
ruff = "^0.1.9"
black = "^23.12.0"
isort = "^5.13.2"
psycopg2 = "^2.9.9"
drf-yasg = "^1.21.7"
django-cors-headers = "^4.3.1"
djoser = "^2.2.2"
django-filter = "^23.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
